+++ frontend/src/components/routes/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { useAuth } from '../../auth/AuthContext';
+import { useAuth } from '../../auth/useAuth';
 import { ROUTES } from '../../app/routes';
 import type { AppRole } from '../../auth/auth';
