+++ frontend/src/components/routes/PublicOnlyRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../../auth/AuthContext';
+import { useAuth } from '../../auth/useAuth';
 import { ROUTES } from '../../app/routes';
 
 function PublicOnlyRoute() {
