+++ frontend/src/components/workspaces/taskboard/chart/ChartView.tsx
@@ -15,7 +15,7 @@ import {
   Legend,
   ResponsiveContainer,
 } from 'recharts';
-import { useTaskBoard } from '../TaskBoardContext';
+import { useTaskBoard } from '../useTaskBoard';
 
 export default function ChartView() {
   const { tasks, groups, boardConfig } = useTaskBoard();
