+++ frontend/src/components/workspaces/taskboard/table/ColumnHeader.tsx
@@ -17,7 +17,7 @@ import AddIcon from '@mui/icons-material/Add';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 import CheckIcon from '@mui/icons-material/Check';
-import { useTaskBoard } from '../TaskBoardContext';
+import { useTaskBoard } from '../useTaskBoard';
 import type { ColumnDefinition, ColumnType, SelectOption } from '../types';
 
 const generateColumnId = () => `col_${Date.now()}`;
