+++ frontend/src/components/workspaces/taskboard/table/MainTableView.tsx
@@ -20,7 +20,7 @@ import {
   sortableKeyboardCoordinates,
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
-import { useTaskBoard } from '../TaskBoardContext';
+import { useTaskBoard } from '../useTaskBoard';
 import TaskGroup from './TaskGroup';
 import TableToolbar from './TableToolbar';
 import { useState } from 'react';
